store/requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
Django==4.2
django-extensions==3.2.1
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
djlint==1.23.0
EditorConfig==0.12.3
factory-boy==3.2.1
Faker==18.4.0
flake8==6.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
